Add nex part of Go code

+++ benchmark/IExecutionContext.go
@@ -2,7 +2,7 @@ package benchmark
 
 type IExecutionContext interface {
 	//parameters interface{}
-	IncrementCounter(increment int64)
+	IncrementCounter(increment int)
 	SendMessage(message string)
 	ReportError(err error)
 	//isStopped bool
